+++ src/ExcelMcp.Core/Commands/Connection/ConnectionHelpers.cs
@@ -1,5 +1,3 @@
-using Sbroenne.ExcelMcp.ComInterop;
-
 namespace Sbroenne.ExcelMcp.Core.Connections;
 
 /// <summary>
@@ -7,49 +5,6 @@ namespace Sbroenne.ExcelMcp.Core.Connections;
 /// </summary>
 public static class ConnectionHelpers
 {
-    /// <summary>
-    /// Gets all connection names from a workbook
-    /// </summary>
-    /// <param name="workbook">Excel workbook COM object</param>
-    /// <returns>List of connection names</returns>
-    public static List<string> GetConnectionNames(dynamic workbook)
-    {
-        var names = new List<string>();
-        dynamic connections = null!;
-
-        try
-        {
-            connections = workbook.Connections;
-            for (int i = 1; i <= connections.Count; i++)
-            {
-                dynamic conn = null!;
-                try
-                {
-                    conn = connections.Item(i);
-                    string name = conn.Name?.ToString() ?? "";
-                    if (!string.IsNullOrWhiteSpace(name))
-                    {
-                        names.Add(name);
-                    }
-                }
-                finally
-                {
-                    ComUtilities.Release(ref conn!);
-                }
-            }
-        }
-        catch
-        {
-            // Return empty list if any error occurs
-        }
-        finally
-        {
-            ComUtilities.Release(ref connections!);
-        }
-
-        return names;
-    }
-
     /// <summary>
     /// Gets the connection type name from XlConnectionType enum value
     /// Per Microsoft docs: https://learn.microsoft.com/en-us/office/vba/api/excel.xlconnectiontype
@@ -71,52 +26,5 @@ public static class ConnectionHelpers
             9 => "NOSOURCE",  // xlConnectionTypeNOSOURCE
             _ => $"Unknown ({connectionType})"
         };
-    }
-
-
-
-    /// <summary>
-    /// Removes connections associated with a query or connection name
-    /// </summary>
-    /// <param name="workbook">Excel workbook COM object</param>
-    /// <param name="name">Name of the query or connection</param>
-    public static void RemoveConnections(dynamic workbook, string name)
-    {
-        dynamic connections = null!;
-
-        try
-        {
-            connections = workbook.Connections;
-
-            // Iterate backwards to safely delete items
-            for (int i = connections.Count; i >= 1; i--)
-            {
-                dynamic conn = null!;
-                try
-                {
-                    conn = connections.Item(i);
-                    string connName = conn.Name?.ToString() ?? "";
-
-                    // Match exact name or "Query - Name" pattern
-                    if (connName.Equals(name, StringComparison.OrdinalIgnoreCase) ||
-                        connName.Equals($"Query - {name}", StringComparison.OrdinalIgnoreCase))
-                    {
-                        conn.Delete();
-                    }
-                }
-                finally
-                {
-                    ComUtilities.Release(ref conn!);
-                }
-            }
-        }
-        catch
-        {
-            // Ignore errors when removing connections - they may not exist
-        }
-        finally
-        {
-            ComUtilities.Release(ref connections!);
-        }
     }
 }
