+++ src/ExcelMcp.Core/Commands/PivotTable/PivotTableCommands.Grouping.cs
@@ -1,4 +1,3 @@
-using Sbroenne.ExcelMcp.ComInterop;
 using Sbroenne.ExcelMcp.ComInterop.Session;
 using Sbroenne.ExcelMcp.Core.Models;
 
@@ -14,53 +13,14 @@ public partial class PivotTableCommands
     /// </summary>
     public PivotFieldResult GroupByDate(IExcelBatch batch, string pivotTableName,
         string fieldName, DateGroupingInterval interval)
-    {
-        return batch.Execute((ctx, ct) =>
-        {
-            dynamic? pivot = null;
-
-            try
-            {
-                pivot = FindPivotTable(ctx.Book, pivotTableName);
-
-                // Determine strategy (OLAP vs Regular)
-                var strategy = PivotTableFieldStrategyFactory.GetStrategy(pivot);
-
-                // Delegate to strategy with logger
-                return strategy.GroupByDate(pivot, fieldName, interval, batch.WorkbookPath, batch.Logger);
-            }
-            finally
-            {
-                ComUtilities.Release(ref pivot);
-            }
-        });
-    }
+        => ExecuteWithStrategy<PivotFieldResult>(batch, pivotTableName,
+            (strategy, pivot) => strategy.GroupByDate(pivot, fieldName, interval, batch.WorkbookPath, batch.Logger));
 
     /// <summary>
     /// Groups a numeric field by the specified interval (e.g., 0-100, 100-200, 200-300)
     /// </summary>
     public PivotFieldResult GroupByNumeric(IExcelBatch batch, string pivotTableName,
         string fieldName, double? start, double? endValue, double intervalSize)
-    {
-        return batch.Execute((ctx, ct) =>
-        {
-            dynamic? pivot = null;
-
-            try
-            {
-                pivot = FindPivotTable(ctx.Book, pivotTableName);
-
-                // Determine strategy (OLAP vs Regular)
-                var strategy = PivotTableFieldStrategyFactory.GetStrategy(pivot);
-
-                // Delegate to strategy with logger
-                return strategy.GroupByNumeric(pivot, fieldName, start, endValue, intervalSize, batch.WorkbookPath, batch.Logger);
-            }
-
-            finally
-            {
-                ComUtilities.Release(ref pivot);
-            }
-        });
-    }
+        => ExecuteWithStrategy<PivotFieldResult>(batch, pivotTableName,
+            (strategy, pivot) => strategy.GroupByNumeric(pivot, fieldName, start, endValue, intervalSize, batch.WorkbookPath, batch.Logger));
 }
