+++ src/ExcelMcp.Core/Commands/Range/RangeCommands.Formatting.cs
@@ -176,14 +176,14 @@ public partial class RangeCommands
                     if (bold != null) font.Bold = bold.Value;
                     if (italic != null) font.Italic = italic.Value;
                     if (underline != null) font.Underline = underline.Value ? 2 : -4142; // xlUnderlineStyleSingle : xlUnderlineStyleNone
-                    if (fontColor != null) font.Color = ParseColor(fontColor);
+                    if (fontColor != null) font.Color = FormattingHelpers.ParseColor(fontColor);
                 }
 
                 // Apply fill color
                 if (fillColor != null)
                 {
                     interior = range.Interior;
-                    interior.Color = ParseColor(fillColor);
+                    interior.Color = FormattingHelpers.ParseColor(fillColor);
                 }
 
                 // Apply borders
@@ -196,8 +196,8 @@ public partial class RangeCommands
                         try
                         {
                             border = range.Borders.Item(edge);
-                            if (borderStyle != null) border.LineStyle = ParseBorderStyle(borderStyle);
-                            if (borderColor != null) border.Color = ParseColor(borderColor);
+                            if (borderStyle != null) border.LineStyle = FormattingHelpers.ParseBorderStyle(borderStyle);
+                            if (borderColor != null) border.Color = FormattingHelpers.ParseColor(borderColor);
                             if (borderWeight != null) border.Weight = ParseBorderWeight(borderWeight);
                         }
                         finally
@@ -240,39 +240,6 @@ public partial class RangeCommands
                 ComUtilities.Release(ref sheet!);
             }
         });
-    }
-
-    private static int ParseColor(string color)
-    {
-        // Support #RRGGBB format or color index
-        if (color.StartsWith('#') && color.Length == 7)
-        {
-            var r = Convert.ToInt32(color.Substring(1, 2), 16);
-            var g = Convert.ToInt32(color.Substring(3, 2), 16);
-            var b = Convert.ToInt32(color.Substring(5, 2), 16);
-            return r + (g << 8) + (b << 16); // Excel RGB format
-        }
-        else if (int.TryParse(color, out var index))
-        {
-            return index;
-        }
-        throw new ArgumentException($"Invalid color format: {color}. Use #RRGGBB or color index.");
-    }
-
-    private static int ParseBorderStyle(string style)
-    {
-        return style.ToLowerInvariant() switch
-        {
-            "none" => -4142, // xlNone
-            "continuous" => 1, // xlContinuous
-            "dash" => -4115, // xlDash
-            "dashdot" => 4, // xlDashDot
-            "dashdotdot" => 5, // xlDashDotDot
-            "dot" => -4118, // xlDot
-            "double" => -4119, // xlDouble
-            "slantdashdot" => 13, // xlSlantDashDot
-            _ => throw new ArgumentException($"Invalid border style: {style}")
-        };
     }
 
     private static int ParseBorderWeight(string weight)
