+++ src/ExcelMcp.Core/Commands/Table/TableCommands.Filters.cs
@@ -22,10 +22,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 // Find column index
                 int columnIndex = -1;
@@ -111,10 +107,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 // Find column index
                 int columnIndex = -1;
@@ -200,10 +192,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 autoFilter = table.AutoFilter;
                 if (autoFilter != null)
@@ -235,10 +223,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 autoFilter = table.AutoFilter;
                 if (autoFilter == null)
