+++ src/ExcelMcp.Core/Commands/Table/TableCommands.Lifecycle.cs
@@ -198,10 +198,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 // Check if new name already exists
                 if (TableExists(ctx.Book, newName))
@@ -232,10 +228,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 // SECURITY FIX: Store range info before Unlist() for proper cleanup
                 try
@@ -280,10 +272,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 sheet = table.Parent;
                 string sheetName = sheet.Name;
