+++ src/ExcelMcp.Core/Commands/Table/TableCommands.NumberFormat.cs
@@ -70,10 +70,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 // Find the column
                 column = FindColumn(table, columnName);
@@ -118,10 +114,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 // Find the column
                 column = FindColumn(table, columnName);
