+++ src/ExcelMcp.Core/Commands/Table/TableCommands.StructuredReferences.cs
@@ -28,10 +28,6 @@ public partial class TableCommands
             try
             {
                 table = FindTable(ctx.Book, tableName);
-                if (table == null)
-                {
-                    throw new InvalidOperationException($"Table '{tableName}' not found.");
-                }
 
                 result.TableName = tableName;
                 result.Region = region;
