--- src/ExcelMcp.McpServer/Prompts/Content/Completions/border_styles.md
@@ -1,9 +1,0 @@
-# Border Styles
-
-none
-continuous
-dash
-dot
-double
-dashDot
-dashDotDot
