--- src/ExcelMcp.McpServer/Prompts/Content/Completions/border_weights.md
@@ -1,6 +1,0 @@
-# Border Weights
-
-hairline
-thin
-medium
-thick
