--- src/ExcelMcp.McpServer/Prompts/Content/Completions/load_destination.md
@@ -1,6 +1,0 @@
-# Load Destination Values
-
-worksheet
-data-model
-both
-connection-only
