--- src/ExcelMcp.McpServer/Prompts/Content/Completions/powerquery_targetCellAddress.md
@@ -1,9 +1,0 @@
-# Completions for targetCellAddress
-
-A1
-B2
-C5
-D10
-E20
-AA3
-AZ25
