--- src/ExcelMcp.McpServer/Prompts/Content/Completions/privacy_level.md
@@ -1,6 +1,0 @@
-# Privacy Levels
-
-None
-Private
-Organizational
-Public
