--- src/ExcelMcp.McpServer/Prompts/Content/Completions/style_names.md
@@ -1,26 +1,0 @@
-# Completions for styleName parameter
-
-Normal
-Heading 1
-Heading 2
-Heading 3
-Heading 4
-Title
-Total
-Input
-Output
-Calculation
-Good
-Bad
-Neutral
-Accent1
-Accent2
-Accent3
-Accent4
-Accent5
-Accent6
-Currency
-Percent
-Comma
-Note
-Warning
