--- src/ExcelMcp.McpServer/Prompts/Content/Completions/validation_operators.md
@@ -1,10 +1,0 @@
-# Validation Operators
-
-between
-notBetween
-equal
-notEqual
-greaterThan
-lessThan
-greaterThanOrEqual
-lessThanOrEqual
