--- src/ExcelMcp.McpServer/Prompts/Content/Completions/validation_types.md
@@ -1,9 +1,0 @@
-# Validation Types
-
-list
-decimal
-whole
-date
-time
-textLength
-custom
